requirements.txt
streamlit
openai
pandas
openpyxl
pypdf
python-dotenv
pillow
python-docx
tiktoken
